+++ lib/src/implementations/scopes.enum.d.ts
@@ -6,5 +6,9 @@ export declare enum CheckdigitScopes {
     PlaidItem = "PlaidItem",
     Supplier = "Supplier",
     DivisionRelationship = "DivisionRelationship",
-    TermsChoice = "TermsChoice"
+    TermsChoice = "TermsChoice",
+    ShopifyCredential = "ShopifyCredential",
+    WoocommerceCredential = "WoocommerceCredential",
+    BigcommerceCredential = "BigcommerceCredential",
+    TradegeckoCredential = "TradegeckoCredential"
 }