+++ src/implementations/checkdigit/historical.scopes.ts
@@ -4,10 +4,3 @@ export enum AlphaHashidScopes {
   Shipment = 'Shipment',
   User = 'User'
 }
-// scopes that were registered in the first version of Arbiter; they do not have checkdigits
-export enum NoCheckdigitArbiterScopes {
-  Workflow = 'Workflow',
-  Activity = 'Activity',
-  BusinessApplication = 'BusinessApplication',
-  ExternalEvent = 'ExternalEvent'
-}