+++ lib/implementations/scopes.enum.d.ts
@@ -1,5 +1,4 @@
 export * from './checkdigit/historical.scopes';
-export * from './banking/banking.scopes';
 export declare enum CheckdigitScopes {
     Buyer = "Buyer",
     Company = "Company",
@@ -18,10 +17,10 @@ export declare enum CheckdigitScopes {
     PendingEvent = "PendingEvent",
     BuyerApplication = "BuyerApplication",
     AchFile = "AchFile",
-    NewAchGateway = "NewAchGateway",
-    NewBankAccount = "NewBankAccount",
-    NewBankTransaction = "NewBankTransaction",
-    NewBatch = "NewBatch",
-    NewOrderReference = "NewOrderReference",
-    NewPaymentRequest = "NewPaymentRequest"
+    AchGateway = "AchGateway",
+    BankAccount = "BankAccount",
+    BankTransaction = "BankTransaction",
+    Batch = "Batch",
+    OrderReference = "OrderReference",
+    PaymentRequest = "PaymentRequest"
 }