+++ lib/implementations/types.d.ts
@@ -1,5 +1,4 @@
 export declare enum ScopeTypes {
-    BANKING = "BANKING",
     CHECKDIGIT = "CHECKDIGIT",
     EXPERIMENTAL = "EXPERIMENTAL"
 }