+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 
 ### Added
+  * `AchFile` oid; shortcode: `ach_`
 ### Removed
 ### Changed
 ### Fixed